+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initialized the back end for the demo application
 - Connected backend to bitcoind
 - Call bridge WS when peg-in funds are deposited
+- Added a "View Wallet" page on the front end